@@ -3,14 +3,8 @@ import SwiftUI
 extension EnvironmentValues {
 	/// - SeeAlso:
 	///   - ``SwiftUI/Scene/aboutWindowLayout(_:)``
-	fileprivate(set) var aboutWindowLayout: AboutWindowLayout {
-		get { self[__Key_aboutWindowLayout.self] }
-		set { self[__Key_aboutWindowLayout.self] = newValue }
-	}
-
-	private struct __Key_aboutWindowLayout: EnvironmentKey {
-		static let defaultValue: AboutWindowLayout = .vertical
-	}
+	@Entry
+	fileprivate(set) var aboutWindowLayout: AboutWindowLayout = .vertical
 }
 
 // MARK: - Convenience
@@ -20,7 +14,7 @@ public extension Scene {
 	/// - Parameter layout: The layout to set in the environment.
 	/// - SeeAlso:
 	///   - ``AboutWindowLayout``
-	func aboutWindowLayout(_ layout: AboutWindowLayout) -> some Scene {
+	nonisolated func aboutWindowLayout(_ layout: AboutWindowLayout) -> some Scene {
 		environment(\.aboutWindowLayout, layout)
 	}
 }
