@@ -3,14 +3,8 @@ import SwiftUI
 extension EnvironmentValues {
 	/// - SeeAlso:
 	///   - ``SwiftUI/Scene/aboutWindowOptions(_:)``
-	fileprivate(set) var aboutWindowOptions: AboutWindowOptions {
-		get { self[__Key_aboutWindowOptions.self] }
-		set { self[__Key_aboutWindowOptions.self] = newValue }
-	}
-
-	private struct __Key_aboutWindowOptions: EnvironmentKey {
-		static let defaultValue: AboutWindowOptions = .default
-	}
+	@Entry
+	fileprivate(set) var aboutWindowOptions: AboutWindowOptions = .default
 }
 
 // MARK: - Convenience
@@ -20,7 +14,7 @@ public extension Scene {
 	/// - Parameter options: The options to set in the environment.
 	/// - SeeAlso:
 	///   - ``AboutWindowOptions``
-	func aboutWindowOptions(_ options: AboutWindowOptions) -> some Scene {
+	nonisolated func aboutWindowOptions(_ options: AboutWindowOptions) -> some Scene {
 		environment(\.aboutWindowOptions, options)
 	}
 }
