@@ -2,7 +2,9 @@ import _MW_Common
 import OSLog
 import SwiftUI
 
-struct ContentView<Content: View>: View {
+struct ContentView<Content>: View where
+	Content: View
+{
 	@Environment(\.aboutWindowLayout) private var aboutWindowLayout
 
 	private let content: () -> Content
@@ -12,14 +14,21 @@ struct ContentView<Content: View>: View {
 	}
 
 	public var body: some View {
-		Group {
-			switch aboutWindowLayout {
-				case .custom: content()
-				case let .vertical(spacing): VerticalContentView(spacing: spacing, content: content)
-			}
+		makePrimaryContent()
+			.onWindowAppear(perform: applyWindowStyle)
+			.windowButtons(miniaturize: .hidden, zoom: .hidden)
+	}
+}
+
+// MARK: - Supporting Views
+
+private extension ContentView {
+	@ViewBuilder
+	func makePrimaryContent() -> some View {
+		switch aboutWindowLayout {
+			case .custom: content()
+			case let .vertical(spacing): VerticalContentView(spacing: spacing, content: content)
 		}
-		.onWindowAppear(perform: applyWindowStyle)
-		.windowButtons(miniaturize: .hidden, zoom: .hidden)
 	}
 }
 
