@@ -18,9 +18,9 @@
 ///   - ``AppVersionView``
 @frozen
 public struct AppVersionOptions: OptionSet {
-	public let rawValue: UInt8
+	public let rawValue: UInt64
 
-	public init(rawValue: UInt8) {
+	public init(rawValue: RawValue) {
 		self.rawValue = rawValue
 	}
 }
