@@ -4,14 +4,8 @@ extension EnvironmentValues {
 	/// - SeeAlso:
 	///   - ``SwiftUI/View/appIconOptions(_:)``
 	///   - ``SwiftUI/Scene/appIconOptions(_:)``
-	fileprivate(set) var appIconOptions: AppIconOptions {
-		get { self[__Key_appIconOptions.self] }
-		set { self[__Key_appIconOptions.self] = newValue }
-	}
-
-	private enum __Key_appIconOptions: EnvironmentKey {
-		static let defaultValue: AppIconOptions = .default
-	}
+	@Entry
+	fileprivate(set) var appIconOptions: AppIconOptions = .default
 }
 
 // MARK: - Convenience
@@ -22,7 +16,7 @@ public extension View {
 	/// - SeeAlso:
 	///   - ``AppIconOptions``
 	///   - ``AppIconView``
-	func appIconOptions(_ options: AppIconOptions) -> some View {
+	nonisolated func appIconOptions(_ options: AppIconOptions) -> some View {
 		environment(\.appIconOptions, options)
 	}
 }
@@ -33,7 +27,7 @@ public extension Scene {
 	/// - SeeAlso:
 	///   - ``AppIconOptions``
 	///   - ``AppIconView``
-	func appIconOptions(_ options: AppIconOptions) -> some Scene {
+	nonisolated func appIconOptions(_ options: AppIconOptions) -> some Scene {
 		environment(\.appIconOptions, options)
 	}
 }
