@@ -3,14 +3,8 @@ import SwiftUI
 extension EnvironmentValues {
 	/// - SeeAlso:
 	///   - ``SwiftUI/View/appInfoSectionStyle(_:)``
-	fileprivate(set) var appInfoSectionStyle: AnyAppInfoSectionStyle {
-		get { self[__Key_appInfoSectionStyle.self] }
-		set { self[__Key_appInfoSectionStyle.self] = newValue }
-	}
-
-	private enum __Key_appInfoSectionStyle: EnvironmentKey {
-		static let defaultValue: AnyAppInfoSectionStyle = AnyAppInfoSectionStyle(.default)
-	}
+	@Entry
+	fileprivate(set) var appInfoSectionStyle: AnyAppInfoSectionStyle = AnyAppInfoSectionStyle(.default)
 }
 
 // MARK: - Convenience
@@ -19,7 +13,9 @@ public extension View {
 	/// Set the style used by ``AppInfoSection``.
 	/// - SeeAlso:
 	///   - ``AppInfoSectionStyle``
-	func appInfoSectionStyle(_ style: some AppInfoSectionStyle) -> some View {
+	nonisolated func appInfoSectionStyle<S>(_ style: S) -> some View where
+		S: AppInfoSectionStyle
+	{
 		environment(\.appInfoSectionStyle, AnyAppInfoSectionStyle(style))
 	}
 }
