@@ -4,14 +4,8 @@ extension EnvironmentValues {
 	/// - SeeAlso:
 	///   - ``SwiftUI/View/appVersionOptions(_:)``
 	///   - ``SwiftUI/Scene/appVersionOptions(_:)``
-	fileprivate(set) var appVersionOptions: AppVersionOptions {
-		get { self[__Key_appVersionOptions.self] }
-		set { self[__Key_appVersionOptions.self] = newValue }
-	}
-
-	private enum __Key_appVersionOptions: EnvironmentKey {
-		static let defaultValue: AppVersionOptions = .default
-	}
+	@Entry
+	fileprivate(set) var appVersionOptions: AppVersionOptions = .default
 }
 
 // MARK: - Convenience
@@ -22,7 +16,7 @@ public extension View {
 	/// - SeeAlso:
 	///   - ``AppVersionOptions``
 	///   - ``AppVersionView``
-	func appVersionOptions(_ options: AppVersionOptions) -> some View {
+	nonisolated func appVersionOptions(_ options: AppVersionOptions) -> some View {
 		environment(\.appVersionOptions, options)
 	}
 }
@@ -33,7 +27,7 @@ public extension Scene {
 	/// - SeeAlso:
 	///   - ``AppVersionOptions``
 	///   - ``AppVersionView``
-	func appVersionOptions(_ options: AppVersionOptions) -> some Scene {
+	nonisolated func appVersionOptions(_ options: AppVersionOptions) -> some Scene {
 		environment(\.appVersionOptions, options)
 	}
 }
