@@ -1,16 +1,11 @@
-import AppKit
+import Foundation
 
-public extension NSApplication {
-	/// The app's bundle identifier.
-	var bundleIdentifier: String! {
-		Bundle.main.bundleIdentifier
-	}
-
+public extension Bundle {
 	/// The app's name.
 	///
 	/// This property reads the value for the key `CFBundleName` in `Info.plist`.  If the key cannot be found, the process name will be used instead.
 	var bundleName: String {
-		Bundle.main.object(forInfoDictionaryKey: kCFBundleNameKey as String) as? String
+		object(forInfoDictionaryKey: kCFBundleNameKey as String) as? String
 			?? ProcessInfo.processInfo.processName
 	}
 
@@ -18,14 +13,14 @@ public extension NSApplication {
 	///
 	/// This property reads the value for the key `NSHumanReadableCopyright` in `Info.plist`.
 	var copyright: String? {
-		Bundle.main.object(forInfoDictionaryKey: "NSHumanReadableCopyright") as? String
+		object(forInfoDictionaryKey: "NSHumanReadableCopyright") as? String
 	}
 
 	/// The app's bundle version.
 	///
 	/// This property reads the value for the key `CFBundleVersion` in `Info.plist`.
 	var bundleVersion: String? {
-		Bundle.main.object(forInfoDictionaryKey: kCFBundleVersionKey as String) as? String
+		object(forInfoDictionaryKey: kCFBundleVersionKey as String) as? String
 	}
 
 	/// The app's version number, optionally including the build number.
@@ -33,7 +28,7 @@ public extension NSApplication {
 	/// The primary app version is retrieved from the value for the key `CFBundleShortVersionString` in `Info.plist`.
 	/// - Parameter includeBundleVersion: If `true`, append the app's build number, retrieved from ``includeBundleVersion``, surrounded by parentheses.
 	func appVersion(includeBundleVersion: Bool = true) -> String? {
-		guard let version = Bundle.main.object(forInfoDictionaryKey: "CFBundleShortVersionString") as? String else {
+		guard let version = object(forInfoDictionaryKey: "CFBundleShortVersionString") as? String else {
 			return nil
 		}
 
