@@ -1,10 +1,12 @@
 import SwiftUI
 
 struct AnyAppInfoSectionStyle: AppInfoSectionStyle {
-	private var _makeBody: (Configuration) -> AnyView
+	private var _makeBody: @MainActor (Configuration) -> AnyView
 
-	init(_ style: some AppInfoSectionStyle) {
-		_makeBody = { configuration in
+	init<S>(_ style: S) where
+		S: AppInfoSectionStyle
+	{
+		_makeBody = { @MainActor configuration in
 			AnyView(style.makeBody(configuration: configuration))
 		}
 	}
