@@ -19,9 +19,9 @@
 ///   - ``SwiftUI/View/launcherActionOptions(_:)``
 @frozen
 public struct LauncherActionOptions: OptionSet {
-	public let rawValue: UInt8
+	public let rawValue: UInt64
 
-	public init(rawValue: UInt8) {
+	public init(rawValue: RawValue) {
 		self.rawValue = rawValue
 	}
 }
