@@ -21,9 +21,9 @@
 ///   - ``SwiftUI/Scene/launcherWindowOptions(_:)``
 @frozen
 public struct LauncherWindowOptions: OptionSet {
-	public let rawValue: UInt8
+	public let rawValue: UInt64
 
-	public init(rawValue: UInt8) {
+	public init(rawValue: RawValue) {
 		self.rawValue = rawValue
 	}
 }
