@@ -22,9 +22,9 @@
 ///   - ``SwiftUI/Scene/recentItemsOptions(_:)``
 @frozen
 public struct RecentItemsOptions: OptionSet {
-	public let rawValue: UInt8
+	public let rawValue: UInt64
 
-	public init(rawValue: UInt8) {
+	public init(rawValue: RawValue) {
 		self.rawValue = rawValue
 	}
 }
