@@ -4,14 +4,8 @@ extension EnvironmentValues {
 	/// - SeeAlso:
 	///   - ``LauncherActionOptions``
 	///   - ``SwiftUI/View/launcherActionOptions(_:)``
-	fileprivate(set) var launcherActionOptions: LauncherActionOptions {
-		get { self[__Key_launcherActionOptions.self] }
-		set { self[__Key_launcherActionOptions.self] = newValue }
-	}
-
-	private struct __Key_launcherActionOptions: EnvironmentKey {
-		static let defaultValue: LauncherActionOptions = .default
-	}
+	@Entry
+	fileprivate(set) var launcherActionOptions: LauncherActionOptions = .default
 }
 
 // MARK: - Convenience
@@ -21,7 +15,7 @@ public extension View {
 	/// - Parameter options: The options to set in the environment.
 	/// - SeeAlso:
 	///   - ``LauncherActionOptions``
-	func launcherActionOptions(_ options: LauncherActionOptions) -> some View {
+	nonisolated func launcherActionOptions(_ options: LauncherActionOptions) -> some View {
 		environment(\.launcherActionOptions, options)
 	}
 }
