@@ -4,14 +4,8 @@ extension EnvironmentValues {
 	/// - SeeAlso:
 	///   - ``RecentItemsOptions``
 	///   - ``SwiftUI/Scene/recentItemsOptions(_:)``
-	fileprivate(set) var recentItemsOptions: RecentItemsOptions {
-		get { self[__Key_recentItemsOptions.self] }
-		set { self[__Key_recentItemsOptions.self] = newValue }
-	}
-
-	private struct __Key_recentItemsOptions: EnvironmentKey {
-		static let defaultValue: RecentItemsOptions = .default
-	}
+	@Entry
+	fileprivate(set) var recentItemsOptions: RecentItemsOptions = .default
 }
 
 // MARK: - Convenience
@@ -21,7 +15,7 @@ public extension Scene {
 	/// - Parameter options: The options to set in the environment.
 	/// - SeeAlso:
 	///   - ``RecentItemsOptions``
-	func recentItemsOptions(_ options: RecentItemsOptions) -> some Scene {
+	nonisolated func recentItemsOptions(_ options: RecentItemsOptions) -> some Scene {
 		environment(\.recentItemsOptions, options)
 	}
 }
