@@ -3,14 +3,8 @@ import SwiftUI
 extension EnvironmentValues {
 	/// - SeeAlso:
 	///   - ``SwiftUI/Scene/recentItemsSearchFilter(_:)``
-	fileprivate(set) var recentItemsSearchFilter: RecentItemsSearchFilter {
-		get { self[__Key_recentItemsSearchFilter.self] }
-		set { self[__Key_recentItemsSearchFilter.self] = newValue }
-	}
-
-	private struct __Key_recentItemsSearchFilter: EnvironmentKey {
-		static let defaultValue: RecentItemsSearchFilter = .default
-	}
+	@Entry
+	fileprivate(set) var recentItemsSearchFilter: RecentItemsSearchFilter = .default
 }
 
 // MARK: - Convenience
@@ -18,7 +12,9 @@ extension EnvironmentValues {
 public extension Scene {
 	/// Customize the search filter for the recent items list.
 	/// - Parameter predicate: The filter to use for each item.
-	func recentItemsSearchFilter(_ predicate: @escaping (String, URL) -> Bool) -> some Scene {
+	nonisolated func recentItemsSearchFilter(
+		_ predicate: @escaping @Sendable (String, URL) -> Bool
+	) -> some Scene {
 		environment(\.recentItemsSearchFilter, RecentItemsSearchFilter(predicate))
 	}
 }
