@@ -1,7 +1,7 @@
 import SwiftUI
 
 extension Text {
-	static func defaultLauncherWindowText() -> Self {
-		Self("WINDOW_TITLE_\(NSApplication.shared.bundleName)", bundle: .module)
+	nonisolated static func defaultLauncherWindowText() -> Self {
+		Self("WINDOW_TITLE_\(Bundle.main.bundleName)", bundle: .module)
 	}
 }
